ofu_app/requirements.txt
django==2.0.1
django-jinja==2.4.1
django-rest-framework==0.1.0
#django-rest-auth==0.9.3
#djangorestframework-jwt==1.11.0
#django-allauth==0.35.0
django-analytical==2.4.0
requests==2.18.4
beautifulsoup4==4.6.0
#psycopg2==2.7.3.2
xmltodict==0.11.0
coverage==3.6
django-cors-headers

